Changed a lot, read description for this commit
Finished app for presentation.
Removed the ios solution, as well as the old api solution.
Added in needed permissions and codecs for release version.
Version 1.0.9 of the app is published on the Play Store and ready for download.

App could definitely benefit from some styling in the future, and more functionality by adding the ability to save players from the roster to the phone's contacts, the ability to integrate calendar events with the stock calendar, and for players to add their picture to their roster card and have it show in the home page as well.